xsigo: Change port number from u8 to u32
change port defination from u8 -> u32 to support
more than 255 value, changes required due to following
commit: 4aacc5eec85d ("RDMA: Support more than 255 rdma
ports")

Orabug: 35180168

+++ drivers/infiniband/ulp/xsigo/xscore/xs_compat.h
@@ -258,7 +258,7 @@ static inline void iboe_mac_vlan_to_ll(union ib_gid *gid, u8 *mac, u16 vid)
 #endif /*IB_REV_106_CHECK */
 
 static inline enum rdma_link_layer rdma_port_link_layer(struct ib_device
-							*device, u8 port_num)
+							*device, u32 port_num)
 {
 	return IB_LINK_LAYER_INFINIBAND;
 }
